Encapsulated service functions in a class
